requirements.txt
cryptography==42.0.8
numpy==2.0.0
opencv_python_headless==4.10.0.84
PyYAML==6.0.1
PyYAML==6.0.1
streamlit==1.36.0
streamlit_authenticator==0.3.2
toml==0.10.2

